+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Configuration module
-- HTTP server factory with reconnect feature
+- HTTP express server factory with reconnect feature
+- Application bootstrap
 
 ## [0.2.0] - 2021-06-24